会员用户模块初版提交

+++ xk-sys-learn-framework/xk-sys-learn-biz-message/src/main/java/cn/xk/xcode/entity/po/BizMessagePo.java
@@ -5,9 +5,6 @@ import com.mybatisflex.annotation.KeyType;
 import com.mybatisflex.annotation.Table;
 import java.io.Serializable;
 import java.time.LocalDateTime;
-
-
-import com.mybatisflex.core.keygen.KeyGenerators;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
@@ -32,7 +29,7 @@ public class BizMessagePo implements Serializable {
      * 消息id
      */
     @Id(keyType = KeyType.Auto)
-    private Long id;
+    private Integer id;
 
     /**
      * 消息类型
